requirements.txt
fastapi[standard]
fastrtc[vad]
python-dotenv
openai
